Better fix for DateInput scroll value reset by applying react-popper inside DatePicker component
