+++ packages/react/src/components/dateInput/DateInput.tsx
@@ -1,11 +1,8 @@
 import { format, parse, isValid, subYears, addYears, startOfMonth, endOfMonth } from 'date-fns';
 import React, { useState, useRef, useEffect, useCallback } from 'react';
-import { usePopper } from 'react-popper';
 
 import { IconCalendar } from '../../icons';
-import classNames from '../../utils/classNames';
 import mergeRefWithInternalRef from '../../utils/mergeRefWithInternalRef';
-import { scrollIntoViewIfNeeded } from '../../utils/scrollIntoViewIfNeeded';
 import { TextInput, TextInputProps } from '../textInput';
 import { DatePicker } from './components/datePicker';
 import styles from './DateInput.module.scss';
@@ -83,7 +80,6 @@ export const DateInput = React.forwardRef<HTMLInputElement, DateInputProps>(
     ref?: React.Ref<HTMLInputElement>,
   ) => {
     const dateFormat = 'd.M.yyyy';
-    const pickerWrapperRef = useRef<HTMLDivElement>();
     const inputRef = useRef<HTMLInputElement>();
     const didMount = useRef(false);
     const [inputValue, setInputValue] = useState<string>(providedValue || defaultValue || '');
@@ -115,92 +111,6 @@ export const DateInput = React.forwardRef<HTMLInputElement, DateInputProps>(
         }
       }
     };
-
-    /**
-     * Close datepicker when clicked outside
-     */
-    useEffect(() => {
-      const handleClickOutsideWrapper = (event: MouseEvent) => {
-        if (showPicker === true) {
-          const isOutside = pickerWrapperRef.current && !pickerWrapperRef.current.contains(event.target as Node);
-          const toggleButton = getToggleButton();
-          const isToggleButton =
-            toggleButton && (toggleButton === event.target || toggleButton.contains(event.target as Node));
-          if (isOutside && !isToggleButton) {
-            const focusToggleButton = document.activeElement === null || document.activeElement.tagName === 'BODY';
-            closeDatePicker(focusToggleButton);
-          }
-        }
-      };
-      window.addEventListener('click', handleClickOutsideWrapper);
-      return () => {
-        window.removeEventListener('click', handleClickOutsideWrapper);
-      };
-    });
-
-    /**
-     * Handle tab inside date picker
-     */
-    useEffect(() => {
-      const pickerModalElement = pickerWrapperRef.current;
-      const tabbleEventHandler = (event: KeyboardEvent) => {
-        // Skip if pressed key was not tab
-        const isTab = event.key === 'Tab' || event.keyCode === 9;
-        if (!isTab) {
-          return;
-        }
-        // Get all focusable elements inside the modal
-        const focusableElements = pickerModalElement.querySelectorAll(
-          'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])',
-        );
-        const firstFocusableElement = focusableElements[0];
-        const lastFocusableElement = focusableElements[focusableElements.length - 1];
-
-        // Check if shift key is pressed
-        if (event.shiftKey) {
-          // If currently focused element is the first element, move focus the last one
-          if (document.activeElement === firstFocusableElement) {
-            (lastFocusableElement as HTMLElement).focus();
-            event.preventDefault();
-          }
-        }
-        // If shift is not pressed and currently focused element is the last element,
-        // move focust the first one
-        else if (document.activeElement === lastFocusableElement) {
-          (firstFocusableElement as HTMLElement).focus();
-          event.preventDefault();
-        }
-      };
-      if (pickerModalElement) {
-        pickerModalElement.addEventListener('keydown', tabbleEventHandler);
-      }
-      return () => {
-        if (pickerModalElement) {
-          pickerModalElement.removeEventListener('keydown', tabbleEventHandler);
-        }
-      };
-    }, []);
-
-    /**
-     * Focus the first focusable element inside the datepicker
-     * when opened.
-     */
-    useEffect(() => {
-      let scrollTimeout;
-      if (showPicker === true && pickerWrapperRef.current) {
-        const focusableElements = 'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])';
-        const firstFocusableElement = pickerWrapperRef.current.querySelector(focusableElements);
-        if (firstFocusableElement) {
-          (firstFocusableElement as HTMLElement).focus();
-        }
-        scrollTimeout = setTimeout(() => {
-          scrollIntoViewIfNeeded(pickerWrapperRef.current);
-        }, 30);
-      }
-      return () => {
-        clearTimeout(scrollTimeout);
-      };
-    }, [showPicker]);
 
     /**
      * Handle the date picker open button
@@ -272,38 +182,7 @@ export const DateInput = React.forwardRef<HTMLInputElement, DateInputProps>(
 
     // Get the current value as Date object
     const inputValueAsDate = stringToDate(inputValue);
-
-    // Initialize Popper.js
-    const { styles: datePickerPopperStyles, attributes: datePickerPopperAttributes } = usePopper(
-      inputRef.current,
-      pickerWrapperRef.current,
-      {
-        placement: 'bottom-end',
-        modifiers: [
-          {
-            name: 'offset',
-            options: {
-              offset: [0, 5],
-            },
-          },
-          {
-            name: 'flip',
-            options: {
-              rootBoundary: 'document',
-              fallbackPlacements: ['bottom-start', 'top-end'],
-            },
-          },
-          {
-            name: 'eventListeners',
-            options: {
-              scroll: false,
-              resize: true,
-            },
-          },
-        ],
-      },
-    );
-
+    const toggleButton = getToggleButton();
     return (
       <div lang={language} className={styles.wrapper}>
         <TextInput
@@ -319,32 +198,25 @@ export const DateInput = React.forwardRef<HTMLInputElement, DateInputProps>(
           inputMode="numeric"
         >
           {disableDatePicker === false && (
-            <div
-              ref={pickerWrapperRef}
-              className={classNames(styles.pickerWrapper, showPicker && styles.isVisible)}
-              role="dialog"
-              aria-modal="true"
-              aria-hidden={showPicker ? undefined : true}
-              style={datePickerPopperStyles.popper}
-              {...datePickerPopperAttributes.popper}
-            >
-              <DatePicker
-                language={language}
-                disableConfirmation={disableConfirmation}
-                selected={isValid(inputValueAsDate) ? inputValueAsDate : undefined}
-                initialMonth={initialMonth}
-                onDaySelect={(day) => {
-                  closeDatePicker();
-                  handleInputChange(format(day, dateFormat));
-                }}
-                onCloseButtonClick={() => closeDatePicker()}
-                selectButtonLabel={getSelectButtonLabel()}
-                closeButtonLabel={getCloseButtonLabel()}
-                minDate={minDate && isValid(minDate) ? minDate : startOfMonth(subYears(new Date(), 10))}
-                maxDate={maxDate && isValid(maxDate) ? maxDate : endOfMonth(addYears(new Date(), 10))}
-                isDateDisabledBy={isDateDisabledBy}
-              />
-            </div>
+            <DatePicker
+              language={language}
+              disableConfirmation={disableConfirmation}
+              selected={isValid(inputValueAsDate) ? inputValueAsDate : undefined}
+              initialMonth={initialMonth}
+              onDaySelect={(day) => {
+                closeDatePicker();
+                handleInputChange(format(day, dateFormat));
+              }}
+              onCloseButtonClick={(focusToggleButton) => closeDatePicker(focusToggleButton)}
+              selectButtonLabel={getSelectButtonLabel()}
+              closeButtonLabel={getCloseButtonLabel()}
+              minDate={minDate && isValid(minDate) ? minDate : startOfMonth(subYears(new Date(), 10))}
+              maxDate={maxDate && isValid(maxDate) ? maxDate : endOfMonth(addYears(new Date(), 10))}
+              isDateDisabledBy={isDateDisabledBy}
+              open={showPicker}
+              inputRef={inputRef}
+              toggleButton={toggleButton}
+            />
           )}
         </TextInput>
       </div>
