+++ packages/react/src/components/dateInput/components/datePicker/DatePicker.module.scss
@@ -21,6 +21,30 @@
   }
 }
 
+.pickerWrapper {
+  transition: opacity 150ms;
+  display: flex;
+  justify-content: flex-end;
+  pointer-events: none;
+  opacity: 0;
+  visibility: hidden;
+  will-change: opacity, visibility;
+  z-index: 900;
+  overflow: hidden;
+  height: 0;
+
+  &.isVisible {
+    opacity: 1;
+    visibility: visible;
+    height: auto;
+    overflow: visible;
+
+    > * {
+      pointer-events: auto;
+    }
+  }
+}
+
 .hds-datepicker {
   background: #fff;
   position: relative;
